Issue #12, #13 are now addressed
